Mainly warnings fixes, from behdad.  http://behdad.org/download/Patches/dasher-alphabet-dir.patch and http://behdad.org/download/Patches/dasher-2-12-build-warnings.patch.

+++ Src/DasherCore/Alphabet/Alphabet.cpp
@@ -192,12 +192,12 @@ void CAlphabet::AddControlSymbol(const string NewCharacter, const string Display
 
 void CAlphabet::Trace() const 
 {
-	unsigned int i;
+	int i;
 	DASHER_TRACEOUTPUT("GetNumberSymbols() %d\n",GetNumberSymbols());
 	DASHER_TRACEOUTPUT("GetNumberTextSymbols() %d\n",GetNumberTextSymbols());
 	
 	int iGroup=0;
-	for (i=0;i<m_Characters.size();i++) 
+	for (i=0;i<(int)m_Characters.size();i++) 
 	{
 		if (iGroup < m_iGroups  && i==m_GroupStart[iGroup])
 		{
